# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (Admin and Manager)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Onboard a user (Admin any role; Manager limited to technician/viewer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a team member (Admin only)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: User removed
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot delete the last Admin or yourself
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Change a user's role (Admin only)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /reports:
    get:
      operationId: listComplianceReports
      tags: [reports]
      summary: List compliance reports
      responses:
        "200":
          description: Compliance reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceReport"
    post:
      operationId: createComplianceReport
      tags: [reports]
      summary: Start a compliance report (Admin and Auditor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceReportInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReport"
  /reports/{reportId}:
    get:
      operationId: getComplianceReport
      tags: [reports]
      summary: Get a compliance report
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReport"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateComplianceReport
      tags: [reports]
      summary: Update a report or advance its workflow stage (Admin and Auditor)
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReport"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteComplianceReport
      tags: [reports]
      summary: Delete a compliance report (Admin only)
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "204":
          description: Report removed
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, auditor, manager, technician, viewer]
        invitedBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
      required: [id, email, name, role, invitedBy, createdAt, lastSeenAt]
    UserInput:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          default: ""
        role:
          type: string
          enum: [admin, auditor, manager, technician, viewer]
      required: [email, role]
    UserRoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [admin, auditor, manager, technician, viewer]
      required: [role]
    ComplianceReport:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [in_preparation, ready_for_review, final]
        periodStart:
          type: ["string", "null"]
          format: date
        periodEnd:
          type: ["string", "null"]
          format: date
        summary:
          type: string
        findings:
          type: string
        rootCauseNotes:
          type: string
        metrics:
          type: object
          additionalProperties:
            type: integer
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - title
        - status
        - periodStart
        - periodEnd
        - summary
        - findings
        - rootCauseNotes
        - metrics
        - createdBy
        - createdAt
        - updatedAt
        - closedAt
    ComplianceReportInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        periodStart:
          type: ["string", "null"]
          format: date
        periodEnd:
          type: ["string", "null"]
          format: date
        summary:
          type: string
          default: ""
        findings:
          type: string
          default: ""
        rootCauseNotes:
          type: string
          default: ""
      required: [title]
    ComplianceReportUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        status:
          type: string
          enum: [in_preparation, ready_for_review, final]
        periodStart:
          type: ["string", "null"]
          format: date
        periodEnd:
          type: ["string", "null"]
          format: date
        summary:
          type: string
        findings:
          type: string
        rootCauseNotes:
          type: string
...